# example/docs/swagger.yml
basePath: /
definitions:
...
  godocker.Container:
    properties:
      Command:
        description: comando em execução
        type: string
      Created:
        description: timestamp Unix de criação
        type: integer
      HostConfig:
        allOf:
        - $ref: '#/definitions/godocker.ContainerHostConfigSummary'
        description: resumo da configuração do host
      Id:
        description: ID completo do container (64 hex)
        type: string
      Image:
        description: 'imagem usada (ex: "nginx:latest")'
        type: string
      ImageID:
        description: ID da imagem (sha256:...)
        type: string
      Labels:
        additionalProperties:
          type: string
        description: labels personalizados
        type: object
      Mounts:
        description: volumes e bind mounts
        items:
          $ref: '#/definitions/godocker.MountPoint'
        type: array
      Names:
        description: 'nomes com "/" na frente (ex: ["/meu-app"])'
        items:
          type: string
        type: array
      NetworkSettings:
        allOf:
        - $ref: '#/definitions/godocker.ContainerNetworkSummary'
        description: resumo de redes conectadas
      Ports:
        description: mapeamentos de porta
        items:
          $ref: '#/definitions/godocker.Port'
        type: array
      SizeRootFs:
        description: tamanho total do filesystem (bytes)
        type: integer
      SizeRw:
        description: tamanho da camada gravável (bytes)
        type: integer
      State:
        description: '"running", "exited", "paused", etc.'
        type: string
      Status:
        description: 'texto legível (ex: "Up 2 hours")'
        type: string
    type: object
...
  godocker.ContainerCreateOptions:
    properties:
      Cmd:
        description: 'comando a executar (ex: ["nginx", "-g", "daemon off;"])'
        items:
          type: string
        type: array
      Entrypoint:
        description: substitui o ENTRYPOINT da imagem
        items:
          type: string
        type: array
      Env:
        description: variáveis de ambiente no formato "KEY=valor"
        items:
          type: string
        type: array
      ExposedPorts:
        additionalProperties:
          type: object
        description: 'portas a expor (ex: {"80/tcp": {}})'
        type: object
      HostConfig:
        allOf:
        - $ref: '#/definitions/godocker.HostConfig'
        description: 'runtime: portas, volumes, restart, limites, etc.'
      Hostname:
        description: hostname do container
        type: string
      Image:
        description: 'imagem a usar — obrigatório (ex: "nginx:latest")'
        type: string
      Labels:
        additionalProperties:
          type: string
        description: 'labels para organização (ex: {"app": "web"})'
        type: object
      OpenStdin:
        description: mantém stdin aberto mesmo sem clientes conectados
        type: boolean
      StopSignal:
        description: 'sinal de parada (padrão: "SIGTERM")'
        type: string
      StopTimeout:
        description: segundos de espera antes do SIGKILL
        type: integer
      Tty:
        description: aloca pseudo-TTY (necessário para shells interativos)
        type: boolean
      User:
        description: 'usuário:grupo para o processo (ex: "1000:1000")'
        type: string
      Volumes:
        additionalProperties:
          type: object
        description: 'pontos de montagem declarados (ex: {"/data": {}})'
        type: object
      WorkingDir:
        description: diretório de trabalho dentro do container
        type: string
    type: object
  godocker.ContainerCreateResponse:
    properties:
      Id:
        description: ID gerado pelo Docker (64 caracteres hexadecimais)
        type: string
      Warnings:
        description: avisos não fatais emitidos pelo Docker (geralmente vazio)
        items:
          type: string
        type: array
    type: object
  godocker.ContainerHostConfigSummary:
    properties:
      NetworkMode:
        description: 'modo de rede: "bridge", "host", "none", etc.'
        type: string
    type: object
  godocker.ContainerInspect:
    properties:
      AppArmorProfile:
        description: perfil AppArmor aplicado
        type: string
      Args:
        description: argumentos passados ao executável
        items:
          type: string
        type: array
      Config:
        allOf:
        - $ref: '#/definitions/godocker.ContainerConfig'
        description: configuração da imagem/container
      Created:
        description: data de criação (RFC3339)
        type: string
      Driver:
        description: 'driver de storage (ex: "overlay2")'
        type: string
      ExecIDs:
        description: IDs de exec em andamento
        items:
          type: string
        type: array
      GraphDriver:
        allOf:
        - $ref: '#/definitions/godocker.GraphDriverData'
        description: dados do driver de armazenamento
      HostConfig:
        allOf:
        - $ref: '#/definitions/godocker.HostConfig'
        description: configuração completa de runtime
      HostnamePath:
        description: caminho do arquivo hostname no host
        type: string
      HostsPath:
        description: caminho do /etc/hosts no host
        type: string
      Id:
        description: ID completo (64 hex)
        type: string
      Image:
        description: ID da imagem (sha256:...)
        type: string
      LogPath:
        description: caminho do arquivo de log no host
        type: string
      MountLabel:
        description: label SELinux para mounts
        type: string
      Mounts:
        description: volumes e bind mounts ativos
        items:
          $ref: '#/definitions/godocker.MountPoint'
        type: array
      Name:
        description: 'nome com "/" (ex: "/meu-container")'
        type: string
      NetworkSettings:
        allOf:
        - $ref: '#/definitions/godocker.NetworkSettings'
        description: configuração completa de rede
      Path:
        description: executável do entrypoint
        type: string
      Platform:
        description: 'plataforma (ex: "linux")'
        type: string
      ProcessLabel:
        description: label SELinux para processos
        type: string
      ResolvConfPath:
        description: caminho do resolv.conf no host
        type: string
      RestartCount:
        description: quantas vezes foi reiniciado
        type: integer
      SizeRootFs:
        description: tamanho total do filesystem
        type: integer
      SizeRw:
        description: tamanho da camada gravável
        type: integer
      State:
        allOf:
        - $ref: '#/definitions/godocker.ContainerState'
        description: estado atual (running, exited, etc.)
    type: object
...
  godocker.Image:
    properties:
      Created:
        description: RepoDigests []string          `json:"RepoDigests"`             //
          digests do registry
        type: integer
      Labels:
        additionalProperties:
          type: string
        description: |-
          SharedSize  int64             `json:"SharedSize"`              // tamanho compartilhado com outras imagens
          VirtualSize int64             `json:"VirtualSize,omitempty"`   // tamanho virtual total
        type: object
      RepoTags:
        description: ParentID    string            `json:"ParentId"`                //
          ID da imagem pai (vazio para base)
        items:
          type: string
        type: array
      Size:
        description: tamanho real em bytes
        type: integer
      id:
        description: ID da imagem (sha256:...)
        type: string
    type: object
...
  godocker.Network:
    properties:
      Attachable:
        description: true = containers podem se conectar manualmente
        type: boolean
      ConfigFrom:
        allOf:
        - $ref: '#/definitions/godocker.ConfigReference'
        description: referência de configuração (Swarm)
      ConfigOnly:
        description: true = apenas configuração, sem instância real
        type: boolean
      Containers:
        additionalProperties:
          $ref: '#/definitions/godocker.NetworkContainer'
        description: 'containers conectados: ID → info'
        type: object
      Created:
        description: data de criação (RFC3339)
        type: string
      Driver:
        description: '"bridge", "overlay", "host", "none", etc.'
        type: string
      EnableIPv6:
        description: true se IPv6 está habilitado
        type: boolean
      IPAM:
        allOf:
        - $ref: '#/definitions/godocker.NetworkIPAM'
        description: configuração de gerenciamento de IPs
      Id:
        description: ID único da rede
        type: string
      Ingress:
        description: true = rede de ingress do Swarm
        type: boolean
      Internal:
        description: true = sem acesso externo
        type: boolean
      Labels:
        additionalProperties:
          type: string
        description: labels da rede
        type: object
      Name:
        description: nome da rede
        type: string
      Options:
        additionalProperties:
          type: string
        description: opções do driver de rede
        type: object
      Scope:
        description: '"local" ou "swarm"'
        type: string
    type: object
...
  godocker.Stats:
    properties:
      blkio_stats:
        allOf:
        - $ref: '#/definitions/godocker.BlkioStats'
        description: métricas de I/O de disco
      cpu_stats:
        allOf:
        - $ref: '#/definitions/godocker.CPUStats'
        description: uso de CPU no momento atual
      id:
        description: ID do container
        type: string
      memory_stats:
        allOf:
        - $ref: '#/definitions/godocker.MemoryStats'
        description: uso de memória
      name:
        description: nome do container
        type: string
      networks:
        additionalProperties:
          $ref: '#/definitions/godocker.NetworkStatsEntry'
        description: métricas de rede por interface
        type: object
      num_procs:
        description: número de processos (Windows)
        type: integer
      pids_stats:
        allOf:
        - $ref: '#/definitions/godocker.PidsStats'
        description: contagem de processos/threads
      precpu_stats:
        allOf:
        - $ref: '#/definitions/godocker.CPUStats'
        description: uso de CPU no momento anterior
      preread:
        description: timestamp da leitura anterior
        type: string
      read:
        description: timestamp da leitura atual
        type: string
      storage_stats:
        allOf:
        - $ref: '#/definitions/godocker.StorageStats'
        description: métricas de armazenamento (Windows)
    type: object
...
  godocker.Volume:
    properties:
      CreatedAt:
        description: data de criação (RFC3339)
        type: string
      Driver:
        description: 'driver do volume (ex: "local")'
        type: string
      Labels:
        additionalProperties:
          type: string
        description: labels do volume
        type: object
      Mountpoint:
        description: caminho real no host
        type: string
      Name:
        description: nome do volume
        type: string
      Options:
        additionalProperties:
          type: string
        description: 'opções do driver (ex: type, device)'
        type: object
      Scope:
        description: '"local" ou "global" (swarm)'
        type: string
      Status:
        additionalProperties: {}
        description: status do driver (depende do driver)
        type: object
      UsageData:
        allOf:
        - $ref: '#/definitions/godocker.VolumeUsageData'
        description: dados de uso (nil se não disponível)
    type: object
...
  main.allData:
    properties:
      containers:
        items:
          $ref: '#/definitions/godocker.Container'
        type: array
      daemon:
        $ref: '#/definitions/godocker.DockerInfo'
      images:
        items:
          $ref: '#/definitions/godocker.Image'
        type: array
      networks:
        items:
          $ref: '#/definitions/godocker.Network'
        type: array
      volumes:
        items:
          $ref: '#/definitions/godocker.Volume'
        type: array
    type: object
  main.apiResp:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  main.statusResp:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: API REST para gerenciar recursos Docker.
  title: godocker API
  version: "1.0"
paths:
...
  /all:
    get:
      description: daemon + containers + images + networks + volumes. Útil para dashboards.
      parameters:
      - description: 1 = inclui containers parados
        enum:
        - 0
        - 1
        in: query
        name: all
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/main.allData'
              type: object
        "500":
          description: Docker inacessível
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Todos os recursos em uma requisição
      tags:
      - Utilitários
  /containers:
    get:
      description: Containers em execução. Use ?all=1 para incluir os parados.
      parameters:
      - description: 1 = inclui containers parados
        enum:
        - 0
        - 1
        in: query
        name: all
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/godocker.Container'
                  type: array
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Lista containers
      tags:
      - Containers
    post:
      consumes:
      - application/json
      description: Cria sem iniciar. Use POST /containers/{id}/start para iniciar.
      parameters:
      - description: Nome do container
        in: query
        name: name
        type: string
      - description: Opções de criação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/godocker.ContainerCreateOptions'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/godocker.ContainerCreateResponse'
              type: object
        "400":
          description: body inválido
          schema:
            $ref: '#/definitions/main.apiResp'
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Cria um container
      tags:
      - Containers
  /containers/{id}:
    delete:
      description: Remove permanentemente. ?force=1 remove mesmo estando em execução.
      parameters:
      - description: ID ou nome do container
        in: path
        name: id
        required: true
        type: string
      - description: 1 = remove mesmo rodando
        enum:
        - 0
        - 1
        in: query
        name: force
        type: integer
      - description: 1 = remove volumes anônimos
        enum:
        - 0
        - 1
        in: query
        name: volumes
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/main.statusResp'
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Remove um container
      tags:
      - Containers
    get:
      description: 'Retorna todos os dados do container: config, rede, estado, mounts,
        etc.'
      parameters:
      - description: ID completo, prefixo ou nome do container
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/godocker.ContainerInspect'
              type: object
        "404":
          description: container não encontrado
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Inspect completo de um container
      tags:
      - Containers
  /containers/{id}/logs:
    get:
      description: 'Últimas N linhas de log (stdout + stderr). Padrão: 100 linhas.'
      parameters:
      - description: ID ou nome do container
        in: path
        name: id
        required: true
        type: string
      - description: Número de linhas (padrão 100)
        in: query
        name: tail
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: erro ao buscar logs
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Logs do container
      tags:
      - Containers
  /containers/{id}/restart:
    post:
      description: Equivale a stop + start. O ?t=N aplica-se à fase de stop.
      parameters:
      - description: ID ou nome do container
        in: path
        name: id
        required: true
        type: string
      - description: Segundos antes do SIGKILL no stop
        in: query
        name: t
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/main.statusResp'
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Reinicia um container
      tags:
      - Containers
  /containers/{id}/start:
    post:
      description: Sem erro se já estiver rodando (idempotente).
      parameters:
      - description: ID ou nome do container
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/main.statusResp'
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Inicia um container
      tags:
      - Containers
  /containers/{id}/stats:
    get:
      description: Snapshot de uso de recursos. Só funciona para containers em execução.
      parameters:
      - description: ID ou nome do container
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/godocker.Stats'
              type: object
        "500":
          description: container parado ou erro
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Métricas de CPU, memória, I/O e rede
      tags:
      - Containers
  /containers/{id}/stop:
    post:
      description: Envia SIGTERM e aguarda. Após timeout envia SIGKILL. Retorna sucesso
        se já parado.
      parameters:
      - description: ID ou nome do container
        in: path
        name: id
        required: true
        type: string
      - description: Segundos antes do SIGKILL (padrão 10)
        in: query
        name: t
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/main.statusResp'
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Para um container
      tags:
      - Containers
  /images:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/godocker.Image'
                  type: array
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Lista imagens locais
      tags:
      - Imagens
  /info:
    get:
      description: Versão, SO, número de containers, memória, plugins, etc.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/godocker.DockerInfo'
              type: object
        "500":
          description: Docker inacessível
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Informações do daemon Docker
      tags:
      - Daemon
  /networks:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/godocker.Network'
                  type: array
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Lista redes Docker
      tags:
      - Redes
  /networks/{id}:
    get:
      description: Driver, configuração IPAM e containers conectados.
      parameters:
      - description: ID ou nome da rede
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  $ref: '#/definitions/godocker.Network'
              type: object
        "404":
          description: rede não encontrada
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Inspect completo de uma rede
      tags:
      - Redes
  /volumes:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/main.apiResp'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/godocker.Volume'
                  type: array
              type: object
        "500":
          description: erro Docker
          schema:
            $ref: '#/definitions/main.apiResp'
      summary: Lista volumes Docker
      tags:
      - Volumes
swagger: "2.0"